# openapi-residence-catalog.yaml
openapi: 3.0.3
info:
  title: Résidence Catalog API
  description: >
    API pour plateforme de promotion immobilière — vente de logements
    (studios, T2, T3) en résidence fermée sur catalogue digital, avec
    paiements échelonnés et suivi de chantier en temps réel.
  version: 1.0.0
servers:
  - url: https://api.residence-catalog.tg/v1
    description: Production
  - url: https://staging-api.residence-catalog.tg/v1
    description: Staging
...
paths:
...
  /catalog/projects/{projectId}:
    get:
      tags: [Catalog]
      summary: Détail d'un programme (blocs, équipements communs)
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        "200":
          description: Détail projet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        "404": { description: Projet introuvable }
...
  /catalog/units:
    get:
      tags: [Catalog]
      summary: Recherche d'unités avec filtres (typologie, prix, disponibilité)
      parameters:
        - name: projectId
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [studio, t2, t3, t4, t5, commerce] }
        - name: status
          in: query
          schema: { type: string, enum: [disponible, reserve, vendu, livre] }
        - name: priceMin
          in: query
          schema: { type: number }
        - name: priceMax
          in: query
          schema: { type: number }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Résultats paginés
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/UnitSummary' }
                  total: { type: integer }
                  page: { type: integer }
  /catalog/units/{unitId}:
    get:
      tags: [Catalog]
      summary: Détail d'une unité (plan, prix, statut, photos)
      parameters:
        - $ref: '#/components/parameters/UnitId'
      responses:
        "200":
          description: Détail unité
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Unit' }
        "404": { description: Unité introuvable }
  /catalog/projects/{projectId}/site-plan:
    get:
      tags: [Catalog]
      summary: Plan de masse interactif — image de fond + position/statut de chaque bloc
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        "200":
          description: Plan de masse
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SitePlan' }

  # ============ RESERVATIONS ============
  /reservations:
    post:
      tags: [Reservations]
      summary: Réserver une unité (pose un verrou temporaire 48h)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [unitId]
              properties:
                unitId: { type: string }
      responses:
        "201":
          description: Réservation créée, en attente de l'acompte
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }
        "409": { description: Unité déjà réservée ou verrouillée par un autre acheteur }

    get:
      tags: [Reservations]
      summary: Lister ses propres réservations
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Liste des réservations de l'utilisateur
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Reservation' }
  /reservations/{reservationId}:
    get:
      tags: [Reservations]
      summary: Détail d'une réservation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        "200":
          description: Détail réservation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Reservation' }

    delete:
      tags: [Reservations]
      summary: Annuler une réservation (libère l'unité)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ReservationId'
      responses:
        "204": { description: Réservation annulée }

  # ============ PAYMENTS ============
...
  /construction/blocks/{blockId}/updates:
    get:
      tags: [Construction]
      summary: Historique d'avancement chantier d'un bloc
      parameters:
        - $ref: '#/components/parameters/BlockId'
      responses:
        "200":
          description: Liste des mises à jour chantier
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConstructionUpdate' }

    post:
      tags: [Construction]
      summary: Publier une mise à jour chantier (admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BlockId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                phase: { type: string, enum: [fondations, gros_oeuvre, second_oeuvre, finitions, livre] }
                progressPercent: { type: integer, minimum: 0, maximum: 100 }
                description: { type: string }
                photos:
                  type: array
                  items: { type: string, format: binary }
      responses:
        "201": { description: Mise à jour publiée, notifications déclenchées }
  /construction/units/{unitId}/progress:
    get:
      tags: [Construction]
      summary: Avancement chantier consolidé pour une unité (hérite du bloc)
      parameters:
        - $ref: '#/components/parameters/UnitId'
      responses:
        "200":
          description: Avancement
          content:
            application/json:
              schema:
                type: object
                properties:
                  phase: { type: string }
                  progressPercent: { type: integer }
                  lastUpdate: { type: string, format: date-time }

  # ============ CLIENT PORTAL ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
    BlockId:
      name: blockId
      in: path
      required: true
      schema: { type: string }
    UnitId:
      name: unitId
      in: path
      required: true
      schema: { type: string }
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    ProjectInput:
      type: object
      required: [name, location]
      properties:
        name: { type: string }
        description: { type: string }
        location: { type: string }
        amenities:
          type: array
          items: { type: string, example: "gardiennage" }
        status: { type: string, enum: [brouillon, publie, complet] }
        siteMapImageUrl: { type: string, description: "Image de fond (scan plan de bornage ou rendu) pour le plan de masse interactif" }
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectInput'
        - type: object
          properties:
            id: { type: string }
            blocks:
              type: array
              items: { $ref: '#/components/schemas/Block' }
            createdAt: { type: string, format: date-time }
    BlockInput:
      type: object
      required: [projectId, name, floors]
      properties:
        projectId: { type: string }
        name: { type: string, example: "Bloc A" }
        floors: { type: integer }
        frontage:
          type: string
          enum: [facade_principale, facade_secondaire, interieur_ilot, mitoyen]
          description: "Position dans la résidence — influence la valorisation prix"
        distanceFromEntranceM: { type: integer }
        sitePlanPolygon:
          type: array
          description: "Points du polygone du bloc sur le plan de masse, en % (0-100) de la largeur/hauteur de l'image de fond"
          items:
            type: object
            properties:
              x: { type: number }
              y: { type: number }
    Block:
      allOf:
        - $ref: '#/components/schemas/BlockInput'
        - type: object
          properties:
            id: { type: string }
            constructionPhase: { type: string }
            progressPercent: { type: integer }
    UnitInput:
      type: object
      required: [blockId, type, surface, floor, price]
      properties:
        blockId: { type: string }
        type: { type: string, enum: [studio, t2, t3, t4, t5, commerce] }
        surface: { type: number, example: 45.5 }
        floor: { type: integer }
        price: { type: number }
        planImage: { type: string }
        status: { type: string, enum: [disponible, reserve, vendu, livre] }
        hasStorefront: { type: boolean, description: "Pertinent uniquement pour type=commerce : vitrine sur rue" }
        streetFacing: { type: boolean, description: "Accès direct depuis la voie publique" }
    UnitSummary:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        blockName: { type: string }
        type: { type: string }
        surface: { type: number }
        floor: { type: integer }
        price: { type: number }
        status: { type: string }
        coverImage: { type: string }
    Unit:
      allOf:
        - $ref: '#/components/schemas/UnitInput'
        - type: object
          properties:
            id: { type: string }
            photos:
              type: array
              items: { type: string }
            constructionProgress: { type: integer }
    Reservation:
      type: object
      properties:
        id: { type: string }
        unitId: { type: string }
        userId: { type: string }
        status: { type: string, enum: [en_attente, confirmee, annulee, livree] }
        lockExpiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...
    ConstructionUpdate:
      type: object
      properties:
        id: { type: string }
        blockId: { type: string }
        phase: { type: string }
        progressPercent: { type: integer }
        description: { type: string }
        photos:
          type: array
          items: { type: string }
        publishedAt: { type: string, format: date-time }
...
    SitePlan:
      type: object
      properties:
        projectId: { type: string }
        projectName: { type: string }
        siteMapImageUrl: { type: string, nullable: true }
        blocks:
          type: array
          items:
            type: object
            properties:
              blockId: { type: string }
              blockName: { type: string }
              frontage: { type: string, enum: [facade_principale, facade_secondaire, interieur_ilot, mitoyen] }
              distanceFromEntranceM: { type: integer, nullable: true }
              sitePlanPolygon:
                type: array
                items:
                  type: object
                  properties:
                    x: { type: number }
                    y: { type: number }
              launchStatus:
                type: string
                enum: [en_commercialisation, seuil_atteint, financement_en_cours, en_construction, livre]
              constructionPhase: { type: string }
              totalUnits: { type: integer }
              soldUnits: { type: integer }
              fillRatePercent: { type: integer }
...